requirements.txt
asgiref==3.10.0
Django==5.2.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
greenlet==3.2.4
inflection==0.5.1
packaging==25.0
playwright==1.55.0
psycopg2-binary==2.9.11
pyee==13.0.0
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
